+++ collator/Cargo.toml
@@ -27,6 +27,7 @@ polkadot-validation = { git = "https://github.com/paritytech/polkadot", branch =
 # Cumulus dependencies
 cumulus-consensus = { path = "../consensus" }
 cumulus-runtime = { path = "../runtime" }
+cumulus-primitives = { path = "../primitives" }
 
 # Other dependencies
 log = "0.4.8"
