+++ pallet-parachain-upgrade/Cargo.toml
@@ -17,7 +17,12 @@ std = [
     'sp-runtime/std',
     'sp-io/std',
     'system/std',
+	'cumulus-primitives/std',
 ]
+
+[dependencies]
+# Cumulus dependencies
+cumulus-primitives = { path = "../primitives", default-features = false }
 
 [dependencies.serde]
 features = ['derive']
