+++ primitives/Cargo.toml
@@ -8,11 +8,17 @@ edition = "2018"
 # Substrate dependencies
 sp-inherents = { git = "https://github.com/paritytech/substrate", branch = "cumulus-branch" }
 
+# Polkadot dependencies
+polkadot-parachain = { git = "https://github.com/paritytech/polkadot", branch = "cumulus-branch", default-features = false }
+
 # Other dependencies
+codec = { package = "parity-scale-codec", version = "1.0.5", default-features = false, features = [ "derive" ] }
 impl-trait-for-tuples = "0.1.3"
 
 [features]
 default = [ "std" ]
 std = [
+	"codec/std",
+	"polkadot-parachain/std",
 	"sp-inherents/std",
 ]
