+++ runtime/Cargo.toml
@@ -12,6 +12,9 @@ hash-db = { version = "0.15.2", default-features = false }
 trie-db = { version = "0.20.0", default-features = false }
 hashbrown = "0.6.1"
 
+# Cumulus dependencies
+cumulus-primitives = { path = "../primitives", default-features = false }
+
 # Substrate dependencies
 sp-std = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "cumulus-branch" }
 sp-runtime = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "cumulus-branch" }
@@ -41,6 +44,7 @@ test-client = { package = "cumulus-test-client", path = "../test/client" }
 default = ["std"]
 std = [
 	"codec/std",
+	"cumulus-primitives/std",
 	"sp-std/std",
 	"sp-io/std",
 	"sp-runtime/std",
