+++ runtime/src/lib.rs
@@ -74,50 +74,3 @@ impl<B: BlockT> ParachainBlockData<B> {
 		&self.extrinsics
 	}
 }
-
-
-pub mod validation_function_params {
-	use codec::{Decode, Encode};
-	use parachain::primitives::{RelayChainBlockNumber, ValidationParams};
-	use sp_inherents::InherentIdentifier;
-
-	/// Current validation function parameters.
-	pub const VALIDATION_FUNCTION_PARAMS: &'static [u8] = b":validation_function_params";
-
-	/// Code upgarde (set as appropriate by a pallet).
-	pub const NEW_VALIDATION_CODE: &'static [u8] = b":new_validation_code";
-
-	/// Validation Function Parameters
-	///
-	/// This struct is the subset of [`ValidationParams`](../../polkadot_parachain/primitives/struct.ValidationParams.html)
-	/// which is of interest when upgrading parachain validation functions.
-	#[derive(PartialEq, Eq, Encode, Decode, Clone, Copy, Default)]
-	#[cfg_attr(feature = "std", derive(Debug))]
-	pub struct ValidationFunctionParams {
-		/// The maximum code size permitted, in bytes.
-		pub max_code_size: u32,
-		/// The current relay-chain block number.
-		pub relay_chain_height: RelayChainBlockNumber,
-		/// Whether a code upgrade is allowed or not, and at which height the upgrade
-		/// would be applied after, if so. The parachain logic should apply any upgrade
-		/// issued in this block after the first block
-		/// with `relay_chain_height` at least this value, if `Some`. if `None`, issue
-		/// no upgrade.
-		pub code_upgrade_allowed: Option<RelayChainBlockNumber>,
-	}
-
-	impl ValidationFunctionParams {
-		pub fn new(vp: &ValidationParams) -> ValidationFunctionParams {
-			ValidationFunctionParams {
-				max_code_size: vp.max_code_size,
-				relay_chain_height: vp.relay_chain_height,
-				code_upgrade_allowed: vp.code_upgrade_allowed,
-			}
-		}
-	}
-
-	/// The identifier for the `validation_function_params` inherent.
-	pub const INHERENT_IDENTIFIER: InherentIdentifier = *b"valfunp0";
-	/// The type of the inherent.
-	pub type InherentType = ValidationFunctionParams;
-}
