+++ runtime/src/validate_block/implementation.rs
@@ -31,7 +31,7 @@ use parachain::primitives::{HeadData, ValidationCode, ValidationParams, Validati
 
 use codec::{Decode, Encode};
 
-use crate::validation_function_params::{
+use cumulus_primitives::validation_function_params::{
 	NEW_VALIDATION_CODE,
 	VALIDATION_FUNCTION_PARAMS,
 	ValidationFunctionParams,
